extract configuration struct/impl code from entity
`#[salsa::input]` structs will need similar code
